moved to_markdown to doc_converter
